Begin client-side code :construction:
Still a work in progress but there is an initial glimmer of hope for the client-
side.  The cycle.js code properly requests data from the api and the koa api
properly sends it back.  The next step is using bookshelf queries to flesh out
the controllers.  Then more work needs to be done cycle-side to get the proper
data cycled in and displayed.  After that, some work on the presentation and
it'll be ready for an early alpha release.